add x-api-key for all route

+++ app/api/username/route.ts
@@ -1,8 +1,13 @@
 // app/api/username/route.ts
 import { NextResponse } from "next/server";
 import { connectDB } from "@/lib/mongodb";
+import { checkApiKey } from "@/lib/checkApiKey";
+import { notFound } from "next/navigation";
 
-export async function GET() {
+export async function GET(req: Request) {
+
+   const forbidden = checkApiKey(req);
+  if (forbidden) return notFound(); 
   try {
     const db = await connectDB();
     const username = await db.collection("username").findOne({});
@@ -20,6 +25,9 @@ export async function GET() {
 }
 
 export async function POST(req: Request) {
+
+   const forbidden = checkApiKey(req);
+  if (forbidden) return notFound(); 
   try {
     const { name } = await req.json();
     
@@ -60,7 +68,9 @@ export async function POST(req: Request) {
   }
 }
 
-export async function DELETE() {
+export async function DELETE(req: Request) {
+    const forbidden = checkApiKey(req);
+  if (forbidden) return notFound(); 
   try {
     const db = await connectDB();
     await db.collection("username").deleteOne({});
